tests/wpt/web-platform-tests/html/semantics/document-metadata/interactions-of-styling-and-scripting/dynamic-render-blocking-link-stylesheet-does-not-block-script.html

<!DOCTYPE html>
<title>Script-created render-blocking link stylesheet is not script-blocking</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<script>
const link = document.createElement('link');
link.rel = 'stylesheet';
link.href = 'stylesheet.py?delay=1';
link.blocking = 'render';
document.head.appendChild(link);
</script>
<h1>Some text</h1>
<script>
test(() => {
  assert_false(styleExists("h1 { color: purple; }"),
               'stylesheet should still be pending');
  const h1 = document.querySelector('h1');
  assert_equals(getComputedStyle(h1).color, 'rgb(0, 0, 0)');
});
</script>
